The testing framework referred to Object Page design model. The directory structure of testing will be shown as follows:
TestingFramework/

 |----config

         |----globalparam.py

|----public

       |----common

              |----basepage.py

              |----log.py

              |----mytest.py

               |----publicfunction.py

              |----pyselenium.py

              |----readconfig.py

       |----pages

|----report

       |----log

       |----testreport

|----testcase

|----run.py

Instructions of each file:

globalparam.py: this file encapsulates the method of writing the file to each folder.

Public folder: files in this folder are public modules.

common: files in this folder are some common operations, such as starting the browser.

pages: this folder is used to store page objects of test cases.

basepage.py: this file encapsulated a base page class for page object

log.py: this file is for generating testing logs.

mytest.py: this file encapsulated a base class of start and end for all test cases.

Publicfunction.py: this file encapsulated a function for move screenshots to img folder

Pyselenium.py: this file encapsulates the configuration of the browser driver. The remote browser driver is used here, and the use of the browser can be selected. There are six browsers to choose from. They are Chrome, FireFox, Internet explorer, Edge, opera and phantomjs. This file also encapsulates some basic functions needed for web testing, these functions call the functions that come with python. Therefore, when testing web pages, it is unnecessary to write more operations , just call these functions.

Readconfig.py； read configuration

Testcase: files in this folder are test cases

Run.py: the entrance of the entire program, after testing operation, it will generate html report and log to report and log folder.